Move components to components/utils, Add dp_05_01

+++ video/utils/sized_container.py
@@ -4,8 +4,8 @@ import manim
 class SizedContainer(manim.VGroup):
     def __init__(
             self,
-            width: int,
-            height: int,
+            width: float,
+            height: float,
             mobject: manim.VMobject,
             show_outline: bool = False,
             **kwargs
